Update README.md
Here I have tried to give a brief explanation of Bayesian Regression for the beginners and people on intermediate level .I have tried to explain each and every  keypoint /factor on which the model depends on. I didn't take down the mathematical proof as it would be intimidating for beginners. Please go through it and do let me know if any improvisation/changes are necessary.